requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
SQLAlchemy==2.0.31
psycopg[binary]==3.2.9
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
pydantic==2.8.2
pydantic-settings==2.2.1
alembic==1.13.2
pytest==8.2.2
httpx==0.27.0
python-dotenv==1.0.1
requests==2.32.3
python-multipart==0.0.12
